books_app/requirements.txt
django==4.0
psycopg2==2.9.3
djangorestframework
python-jose == 3.3.0
pydantic==1.9.2
bcrypt